Refactored invariants specification automata
Grouped all ISA implementation in a builder class
InvariantsSpecificationAutomatonBuilder. An ISA can
now be built directly after the original correctness
witness automaton is created and analyzed. Moreover, if
an ISA is specified, it overwrites the witness automaton
so there will be only the ISA or the witness automaton
in the main analysis. The refactoring makes also now the ISA
printable. Refactored WitnessExtractor class to make it more
understandable and to adapt it for the new ISA builder class.
Added JavaDoc for some methods.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/witnessValidationPredicate@31331 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c