Update bundled BenchExec to 1.20-dev0 (b2ea4d4)
This contains a change how tool-info modules are called
(inside a container if --container is used),
and I want to give this a little bit more real-world usage before the
next BenchExec release.
Thus, please use --container for your benchmarks,
even if you use the VerifierCloud.

In this BenchExec version, the auto build of CPAchecker has been removed
from the tool-info module:
https://github.com/sosy-lab/benchexec/commit/b2ea4d4b1ec4101d8f575a06d5e14507bfba81d5

Because we decided in the discussion of #549 to keep the auto build,
I added code to benchmark.py that patches BenchExec and adds the auto
build if CPAchecker is benchmarked.
This means that the auto build is only active if benchmark.py is used
and not if pure benchexec is used, but I guess there are not that many
people who benchmark source checkouts of CPAchecker,
and for them there is still at least the check in the tool-info module
that tells them they should build manually if necessary.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@31341 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c