BAMPredicateReducer: bugfix for assertion in BAMBlockFormulaStrategy#getFormulasForPath.
Without a unique instance of AbstractionFormula,
the merge-operator will merge states from different blocks.
This is against the tree-like shape of PredicateAbstractionStates in the ARG
and BAMBlockFormulaStrategy crashed when re-computing the pathformulae.

By default, the reduction of AbstractionFormulae already produces a new instance.
However, this does not happen if reduction is disabled.

Small example:
scripts/cpa.sh -predicateAnalysis-bam \
    -setprop cpa.predicate.bam.useAbstractionReduction=false \
    -spec test/programs/simple/ALL.prp \
    test/programs/simple/nested_operators/nested-logical-binary-operators_true-unreach-label.c

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@31491 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c