Disable tests with OptiMathSAT on GitLab
We want to enable the ConfigurationFileChecks on modern Java
and need to find the cause of the segfaults first.
Previously they occured in liboptimathsat5j.so,
but now in libmathsat5j.so.
So we disable OptiMathSAT for now to make sure it does not interfere
and see if the segfaults still occur.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@31513 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c