Re-disable tests with OptiMathSAT
It seems indeed to be the case that OptiMathSAT triggers the segfaults
on GitLab CI, even if they occur in libmathsat5j.so.

Since our binary of OptiMathSAT is quite old and needs to be updated
anyway (https://github.com/sosy-lab/java-smt/issues/132),
we just disable the tests until the update was done.

We also disable the upload of core dumps
because they are too large (several GB).

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@31516 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c