BAMPredicateCPA: remove RelevantPredicatesComputer.
The RelevantPredicatesComputer is/was an optimization and a heuristic.
For predicates that are not used inside a block,
this special component removed them from the abstractionFormula before entering (reduce)
and re-added tehm after leaving (expand).

However, this components has several serious issues:
- It works fine for simple predicates.
  Pointers, aliasing and adress support is missing.
- If the AbstractionFormula is no conjunction,
  but disjunctions or more complex operations,
  the re-adding leads to a weaker formula when expanding the block.
  This was partially curcumvented by the "refinable" RelevantPredicatesComputer.
- The components does not match our structure of data- and operator-classes
  and causes a "global mofifiable" state in the CPA,
  including memory leaks in the BAM-cache and
  conflicts with further development like block-local-CEGAR.

We tried to circumvent the first two problems mentioned above
with using the refinable RelevantPredicatesComputer and by repeating refinements.
However, there is no guarantee that this works for all programs
(there are several larger programs (mainly LDV) with undebuggable issues).
The manual effort to track down every case is too large and
the component is just a heuristic, thus we remove it.

The new solution contains parts of the existing code directly in the BAMPredicateReducer
and works demand-driven, i.e. computes the relevant predicates right when needed.
There will be a second commit soon that improves the reduction/expansion.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@31518 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c