BAMPredicateCPA: improve reduction/expansion by only removing implied predicates.
This is related to the last commit.
Previously, we removed all possible predicates from the abstraction.
This leads to imprecise expanded formulae and required special handling
like a refinable set of relevant predicates (which was removed in the last commit).
Now, we only remove predicates that are implied by the abstraction formula,
such that re-adding the removed predicates is possible
and results in a precide abstraction at expansion.

For example, we had removed a relevant predicate P from a disjunctions like "(A & P) || B",
which is not possible now, because the formula does not imply P.

This commit moves the code for reduction/expansion directly into the BAMPredicateReducer
and applies the described changes.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@31519 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c