BAMPredicate: further simplification of logic for repeated cex.
This commit actually removes the code for tracking repeated cex for BAMPredicate.
The code was already unused and possibly faulty,
because it only checked for repeated cex once, i.e.,
repeating twice was allowed (and a third appearence just continued).
There has never been a guarantee that a second iteration
is sufficient to update all aggressively cached entries,
maybe even a third or fourth iteration was needed.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@31648 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c