Refactoring: more efficient transformed copies of list
The common cases of FluentIterable.from(list).transform(...).toList()
and list.stream().map(...).collect(toImmutableList()) are inefficient
because they do not make use of the fact that the result has the same
size as the input.
Best would be to directly allocate an appropriately pre-sized list,
but instead they copy the transformed values into a growing list builder
and then create the final list from that.
Our helper method avoids this unnecessary performance loss.

The same is done for copying into an ImmutableSet.

This commit was automatically created with Refaster.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@31668 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c