Make CPAchecker more deterministic
Even in cases of explicitly "random" strategies we only want some
arbitrary orer, not a nondeterministic one, because we want CPAchecker
to always return the same result in the same time (as far as possible).

We make sure that the Random instances with constant seed are static,
such that we get the same deterministic behavior across CPAchecker runs,
but still different behavior if a class is instantiated several times.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@31674 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c