PointerTargetSet: avoid exponential blowup when joining nearly identical union types.
Problem:
Merging a union type "UNION(A,B)" with a new "A" resulted in "UNION(A,B,A)".
Merging a union type "UNION(A,B)" with a "UNION(A,B)" resulted in "UNION(A,B,A,B)".
The second "A"/"B" is not needed in both cases.
This behaviour is especially bad for repeated PTS merges,
because the overhead (memory and runtime) can grow fast.

Solution:
Lets filter the inner members of (already merged) union types and remove duplicates.
In both examples above, the result "UNION(A,B)" is sufficient.
Additionally, change the type-name to represent the filtered union-type to avoid large Strings.

This change seems to avoid the exponential blowup from #602.

PS:
I do not yet know whether merging PTS on expand operation in BAM is fully correct.
However, this commit might also be useful in other cases when using PTS.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@31680 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c