Simplify and optimize sorted copies of some collections
For example, sorting using toString() is inefficient if we only need the
strings afterwards anyway (because toString() is called often).
We can instead first map to strings and then sort.

Using ImmutableList.sortedCopyOf is more efficient than sorting a list,
because it can directly sort the underlying array.
(Ordering.immutableSortedCopy is the same, but less obvious.)

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@31690 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c