Fix two cases where InvalidConfigurationException was converted into AssertionError
This is just wrong: An AssertionError is for cases that should never
occur except in case of bugs, but a user giving a wrong configuration is
not such a case.
Furthermore, it is a bad user experience if CPAchecker crashes with a
stack trace because a configuration value is wrong.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@31729 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c