Remove AutomatonCPA from AutomatonState
In nearly all cases the automatonCPA was only used to either retrieve
its stored automaton or the name of the automaton.
In the remaining three cases the GlobalInfo instance is now used to get
the automatonCPA from the respective automaton. This needs yet to be
tested though.

This is mainly done for having more flexibility in DCACPA, because
the CPA makes use of multiple automata, however, for now it always
instantiates each of the AutomatonStates with the same
ControlAutomatonCPA (thus being able to contain the wrong automaton).

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/ltl-model-checking@31730 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c