Fix invariants in exported witnesses
Before, too many states in the witness were merged. If a state with
invariant 1(true) gets merged into a state with an actual invariant
like i==0, this will lead to a new state with invariant 1. This is
sound in a way that it does not produce wrong invariants, but it leads
to witnesses without any invariants, e.g. in case of predicate
analysis with ABE, since all non-abstraction states have invariant 1.
This commit will change witnesses from predicate analysis
significantly. Witness validation might now take longer with actual
invariants. k-Induction is probably not affected much by this as it
usually adds the invariants also to neighbouring locations where they
apply, thus probably preventing most of the destructive disjunctions.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@31740 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c