Revert r31740
The change of r31740 was not wrong in general, but this led to
possible compatibility issues with other validators. The complicated
part is the structure of the head of the while loop. We do not want
to have multiple edges in sequence that match the line of the head
of the while loop, and we also do not want to have epsilon edges in
the witness. A different validator is not guaranteed to use the same
CFA structure as we have. More details in #606.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@31764 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c