Benchmarking with VCloud: fix inconsistent output-directory.
Additional output files (witnesses,...) are written into a special directory.
This directory differs between direct and webclient mode when using VCloud.

As there is currently no fix from VCloud-side,
we fix the inconsistency from CPAchecker-side in the scripts,
i.e. simply move the corresponding files into the correct location automatically.

The affected users group includes only direct users of VCloud
and no external users that use the web-interface via http.

For more information on the problem, there is an issue for the VCloud:
https://gitlab.com/sosy-lab/software/verifiercloud/issues/396 .

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@31798 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c