Extract new class for bundling witness information
This is one step further to separating the witness information from the
act of writing it to a file, e.g. as GraphML. Next, the WitnessWriter
will be decoupled from the GraphML-Code completely and only used to
generate Witness objects, probably in a factory method pattern.
This might also be the time to rename the class WitnessWriter.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@31857 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c