Finish refactoring of WitnessWriter
WitnessWriter is renamed to WitnessFactory. It is now possible to
separate the gathering of information that is needed for the witness
(into the new Witness object) from the process of writing it into a
specific format such as GraphML. This will ease transforming the
witness into different formats, such as for visualization in the
HTML report or as annotation in the program itself. It is also
much easier to understand what is going on now.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@31859 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c