Add witness as view of the ARG to the HTML report
The witness is created by taking the ARG and merging some of the nodes
together, e.g. those at the same CFA location. By adding the witness
as special view on the ARG we can sometimes get a more compact form
compared to the simplified ARG. This of course also helps to better
understand the generated witness and why it looks how it looks.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@31934 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c