Created Analysis for usage of Array elements
The analysis is based on Cousot et al. "A parametric segmentation functor for fully automatic and scalable array content analysis" (2011) and the Master Thesis of Jan Haltermann. The current version is a prototypic implementation.
It contains two main components: On the one hand, there is the implementation of the Array Segmentation domain, containing everything needed to lift an analysis for simple variables to arrays. On the other hand, we have a simple analysis to detect usage of array elements, showing how to use the array segmentation domain.
It is assumed, that no “Array-Index-out-of-bound” errors occur in the analysed program.

-----------------------------------------------------------------------

Extended the array segmentation domain and the unification algorithm

Now, we support the following extra features:
- In case that there are branches inside of loops, we avoided merging array segments if not needed by extending the unification algorithm
- We added splitting of array segmentations to be able to use constants inside the array segmentation. We again extended the unification algorithm to make this work.
- Extended the storing of analysis information for array segments to get more precise results
Moreover, we provided an abstract CPA Analysis, making lifting analysis to arrays easier.

Known issues that need to be fixed and planned extensions:

- The analysis may not work correctly if there are branches in loops, where the number of statements in the branches differs
- Currently, we are assuming that the variable used to access the array(‘i’) is always bounded above by the length of the array. Moreover, we assume that the size of the array does not overflow, hence we can assume that ‘i+1-1 = I’ holds (same assumption as in original paper).
- The path formula is currently computed using a slight modified implementation from Manuel Tows. We plan to use the predicate analysis to compute the path formula. (+1 squashed commits)

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/arrayContentAnalysis@32119 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c