Remove manipulation of PATH variable by benchmark.py
Previously, some directory under lib/native/ was added to the PATH.
This was added in 2015 (when BenchExec got separated from CPAchecker)
to keep the previous behavior (that was added in 2011 in r3202 (070ebcdb))
of searching for CBMC binaries in lib/native/ in order to simplify
benchmarking of CBMC.

However, this is bad:
- Execution of CBMC by CPAchecker itself must work without manipulation
of PATH, because otherwise it would work only during benchmarking, not
during manual execution of CPAchecker.
- When users of benchmark.py want to benchmark CBMC, we should not
silently default to the (probably outdated) binary under lib/native/,
but let the users provide their own CBMC version.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@32157 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c