FormulaEncoding: add option to allow lazy loading of theories.
Having an option might be a compromise between
- missing a nice user message at all,
- adding InvalidConfigException to all formula-related methods (there are plenty of them),
- just further throwing the UnsupportedOperationException from JavaSMT.

The new option is by default set such that the nice user message is shown.
For new solvers (hopefully coming in a few weeks),
we might consider to carefully disable the option on demand.
Some solvers miss basic theories like Integer or Rational and
provide no possibility to encode Floats.
The new option allows avoid an exception and to analyze programs
that do not even require encoding for Floats.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@32165 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c