SMG: fix comparison of memory adresses that are no longer valid.
Our implementation currently does not reuse memory adresses after freeing them.
Thus we get new (unequal) Region identifiers.
This commit is a first step, such that we can compare such pointers.

A better solution would be to cleanup the SMG after free,
such that no trace of the used object identifiers remains.
We should think about that in the future.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@32260 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c