fix exception for CProblemTypes in additional assumptions.
Lets ignore assumptions with CProblemTypes
that come from forther sources like a witness automaton,
also as part of the counterexample check after ana analysis.
We print a info for the user about that.

The exception is at least thrown for the following comandline:
  scripts/cpa.sh -svcomp20 -heap 10000M -benchmark -timelimit '900 s' \
    -64 -stats -spec ../sv-benchmarks/c/properties/unreach-call.prp \
    ../sv-benchmarks/c/ldv-linux-3.4-simple/32_7_cilled_const_ok_linux-32_1-drivers--isdn--capi--kernelcapi.ko-ldv_main3_sequence_infinite_withcheck_stateful.cil.out.i
where an assumption '*(g_cmsg->BC) == 0U' is used in the counterexample check
and the type of the left operand (pointer from struct) is unknown.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@32316 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c