integration-sets: remove several tasks that only produce timeouts.
The BuildBot has limited resources and can not execute all possible tasks.
This change will help to improve the roundtrip time for getting results from BuildBot.

Lets keep the tasks on a level where each integration-set can be executed
within one or two hours on a single machine.
This commit removes several blocks of tasks where the result is
mostly 'timeout' or 'failure' without any further usage for the developer.

For some integration-sets (like witness validation or value analysis),
there is already an additional nightly BuildBot job
that runs on the complete ste of sv-benchmarks.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@32336 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c