Switch build system to Java 11
If you experience compile errors, run "ant clean build".

With this commit, CPAchecker can no longer be compiled with Java 8.

We need to replace the annotation javax.annotation.Generated with
javax.annotation.processing.Generated, but this allows us to remove a
hack in the build system where we replaced the annotations
just before and after compilation.

Fixes #591.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@32340 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c