Hard-code default for solver.nonLinearArithmetic for termination analysis
This is a JavaSMT option, but CPAchecker wants to use a different
default than the JavaSMT default.
This was implemented in the class Solver and used by all code except for
the termination analysis. There we relied on setting it in the configs,
but this can be forgotten more easily (cf. #682).
So now we set the default in the code of all analyses that use a solver,
and can remove it from configs.

For old SV-COMP configs we add the option such that the behavior does
not change.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@32376 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c