Explicitly mark ForceTerminationOnShutdown as canceled after analyis completes
During printing of statistics and writing of output files,
ForceTerminationOnShutdown should not kill CPAchecker.
So far we relied on the interrupt signal of the
ForceTerminationOnShutdown thread, but there are cases
where it still triggers as discussed in
https://gitlab.com/sosy-lab/software/cpachecker/issues/562#note_258515191

With the explicit boolean for cancelation
it should be ensured that it never triggers after the analysis ends.

Part of #562.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@32409 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c