BAM: allow Reducer to return null on expanding an unsatisfiable state.
Combining the block abstraction and the intiial state during
the expand operation can lead to an unsatisfiable and unreachable state.
We return NULL in that case.

This might only happen if the reduce operation
does not satisfy the implication rule according to stronged post (limited to block variables).
TODO: We need to enable a repeated CPA-algorithm in such a case, or just look
for all possible exit and target states in the reached-set, i.e., until the waitlist is empty.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@32531 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c