BDD: add a new analysis configuration that uses full block abstraction with BDDs.
The new analysis aims towards analyzing each block exactly once and provides full block abstractions.
Currently, the new analysis is in many tasks slower than the existing BDD analysis (with and without BAM).

The new analysis is basically conceptually different
to the BAM-based approaches in Value or Predicate analysis,
where we keep inductive information for reduce and expand operation.

Until now, the most similar approach in CPAchecker might be BAMrec for Predicate analysis.
However, even there, we track information across the block entry,
and this is not done in the new BDD reduction.

The BDD block abstraction is fully decoupled of the surrounding block.
We use additional variables as input and apply variables:
the first as reference point for building the block abstraction,
the second to avoid name collisions during block abstraction application.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@32556 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c