Convert CompoundStatementExpression to null if its return type is void
Revert changes from #585 because of discussion in #696.
The type of a CompoundStatementExpression of type void is inferred as ProblemType, resulting in the creation of a temp variable.
However, the value of the CSE cannot be assigned to the temp variable, which results in the error from #585.
The changed code checks the type of a CSE and returns null instead of a temp variable, if the CSE is void, which is the expected behaviour.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/695-follow-up-from-585@32685 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c