Refactor ARGToCTranslatorTest: Divide into subclasses
Make one subclass for basic tests of ARGToCTranslator on ARGs,,
one for tests of ARGToCTranslator on ARGs that were created
with special specifications.

This allows a more flexible extension for other combinations,
without the need to extend the list of constructors
for every additional option one would like.

E.g.: I wanted to add tests for ARGs created from
combinations with condition automata. For this, I had to
add a constructor parameter for the corresponding condition automaton.
For this, I would have to touch all existing tests to add the additional
parameter with value 'null'.
This would repeat for any additional test setting.
Now, we can use separate test-classes for every test scenario,
and only use the constructor parameters necessary

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@32776 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c