Make slicing configs always use CEX-check and use CEX from that check
Because counterexamples found on the program slice may be invalid,
use counterexamples deemed valid by the counterexample check.

It is necessary to use the counterexample from the CEX check
because the CEX-check does not check the concrete CEX
added to the target state in the ARG. Instead,
the CEX check only checks whether any of the found target paths
is feasible.
Since this can be true even if the concrete CEX in the ARG
is infeasible, use the CEX computed by the check to always get a valid
one.

Example for when this is necessary:
```
scripts/cpa.sh -predicateAnalysis-slicing \
        -setprop solver.solver=Z3 \
        test/programs/program_slicing/branch_none_relevant2.c
```
(solver=Z3 is necessary because of current MathSAT5 crash)

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@32793 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c