When using CEX from CEX-check, use exchangeable version of ARGPath
An ARGPath is dependent on the ARG it describes a path in,
but the ARG computed by the CEX-check is destroyed after
the CEX-check finishes.
To still provide a valid ARGPath in the used counterexample,
create a new, stripped-down version of the ARGPath.
The ARGStates of that new ARGPath
only consist of the LocationState and the AutomatonStates
of the corresponding original ARGStates.
These stripped-down versions can be safely handed between components.

Since the resulting ARGPaths miss information of the original
computation,
we fully rely on the CFAPathWithAssumptions to provide information
crucial to the CEX.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@32802 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c