Add option to BMC for slicing unreachable states
Our incremental BMC analysis tends to blow up because it keeps
creating successors even when the path formula  to a state at
the boundary is unsatisfiable. Checking this for each state that
will be used to extend the reached set in the next incremental
iteration should speed up things significantly.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@32812 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c