Do not add SSA indices for variables representing calls to __VERIFIER_nondet_int
The predicate analysis creates a fresh variable for all calls to
__VERIFIER_nondet_int and similar functions. So far it uses a constant
name and uses the SSA index to create unique variables.
However, this means that these indices get removed when uninstantiating
a formula, but this is actually not desired.
For example, predicate could not talk about a value that results from a
specific call to __VERIFIER_nondet_int.
Usually problems like this do not appear, because interpolants and
predicates never contain such variables (because they occur only once
in a formula), but with other methods than interpolation this could
happen.

Now we use a variable name that changes and mark the variable as having
no SSA index. We still use the SSAMap to store the counter for the next
free variable name, though. This is necessary because when we recreate
formulas for a path, we need to produce the same formula, and a global
counter stored in the CtoFormulaConverter would violate this
requirement.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/predicate-mark-nondet-variables@32896 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c