Make path created by PathTranslator simpler
If we have an ARGState with two children and AssumEdges leading to them,
we can expect the two edges to be the negation of each other.
By unrolling the loop that handles the two edges we can replace
"if (foo) {} else if (!foo) {}" with "if (foo) {} else {}".
This makes the path easier to read and might also make it easier for
other tools to consume.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@32942 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c