Blacklist varargs implementations using cpa.callstack.unsupportedFunctions
This means that CPAchecker will return UNKNOWN if it encounters a
function that has varargs during the analysis.
This is better than the current state where we are imprecise and not
even counterexample checks can help due to the way we rewrite the
program.

This should mostly affect the BusyBox tasks, which have a reachable but
unimportant varargs function.

Note that this does not affect calls to extern varargs functions.

Part of #711.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@32945 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c