Make cpa.callstack.unsupportedFunctions match only extern functions again
There is no reason why we would want to abort the analysis just because
a function in the program has a specific name.
We only want to abort the analysis if we encounter calls to a specific
extern function of which we know that we do not support its semantics.

This reverts r26443 (47c71485) and r29744 (93c73878).

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@32946 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c