ParallelBAM: bugfix for race condition in scheduling.
__Problem__:
For about 1-2% of several hundred tasks, an exception occured randomly
when shutting down the thread pool after an execution of ParallelBAM.
For example, see the following table in the BuildBot:
https://buildbot.sosy-lab.org/cpachecker/results/bam/00508.bam.diff.html

__Details__:
Whenever a RSE was scheduled several times sequentially,
it can happen that the earlier RSE executions already do all the work
and cleanup all dependencies between the corresponding reached-sets.
The remaining RSE executions are not yet applied (they 'wait' for an execution).
As we close the thread pool after all dependencies are solved,
the CompletableFuture framework crashes when trying to schedule the remaining RSE executions.

__Solution__:
We need to count the number of scheduled jobs and
only close the thread pool if no remaining jobs exist.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@32964 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c