Set option counterexample.export.allowImpreciseCounterexamples to false
This enables an assertion that was added in r29840 (d4203281).
With this assertion we will find problems such as #717 more easily.

We need to disable it for slicing because of #716
and for -ldv because of #283,
this should be removed after the respective issue is fixed.

If there are other analyses for which this option creates too many
problems, the option can be disabled in the respective config files.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@32971 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c