Make popover with ranking-information appear focus-only, right
If the popover appears 'bottom', it is not fully visible
in small windows. Now the popover appears 'right' and may
cover the button that makes it appear, but this doesn't matter:
We also remove the data-toggle 'click' and only use 'focus' instead.
There is no reason for anyone to have two of these big popovers open
at the same time, since the detailed fault-view is better-suited
for comparisons. So let's avoid them clutter the error path.
In addition, I claim that 'unfocusing' is the default mechanism
to make popovers disappear on the web.

This also gets rid of a bug: The data-toggle attribute
was specified twice, once with 'focus' and once with 'click'.
In my firefox, this lead to bootstrap only using the 'focus' behavior.
The correct way to specify two values for the attribute would
be `data-toggle: 'focus click'`. But this is a disjunction,
so to make a popover disappear one would have to click the button
*and* unfocus it.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/fault-localization-structure@33198 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c