First draft of a portfolio analysis making (rudimentary) use of MPI
For running CPAchecker with the portfolio flag, the user is currently
required to have both python3 and the python package 'mpi4py' installed
on its system (the latter can be installed using e.g. 'pip3 install
mpi4py').

Currently, the algorithm is able to run several CPAchecker instances
using MPI, however, the results are not yet collected. Thus, the
main-node of CPAchecker will always report an unknown result as of
now.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@33390 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c