Remove library that was accidentally added in r33403
For a more comprehensive background:
The spot library is an alternative option for parsing ltl formulaes into
buechi-automatai (instead of e.g. ltl3ba). However, it needs to be recompiled
because of missing internal shared libraries, and as I never had the time to do
that, it remained unused on my local machine.

In my last commit, I removed the mpi-related files using 'git add
lib', thus also accidentally adding the spot binary.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@33405 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c