Cleanup ant targets for releases
Our main archive for Linux binaries is now a ZIP,
because we need that for SV-COMP anyway,
it is more practical when uploading to Zenodo,
and there is no strong argument for TAR.

We keep the target for creating a TAR
for example because it is easier to use in build/Dockerfile.dev.

Also cleanup some deprecated targets.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@33442 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c