Make Report support highlighting of multiple path elements and share state between tabs
When an error-path element is clicked, don't just highlight the element
in the corresponding CFA/ARG **or** source-code tab that is currently
open, but highlight it in all of the tabs.
This way, when a user changes the tab, the same element is highlighted.

Previously, a clicked error-path element was only highlighted in the
current tab and other tabs could have old elements highlighted.

Simplify call structure and get rid of redundant methods in the process.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/kettl-fault-localization@33479 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c