Use correct raw signature in CFA for conditions
We don't have a way to use the correct signature
for switch-case statements, at the moment, so let's
add a FIXME to that.

The incorrect raw signature is an issue when exporting
condition automata or violation witnesses.

Example to reproduce previous issue on branches:

  Task: https://github.com/sosy-lab/sv-benchmarks/blob/6c4d8bc9d7cecadf3c4a4bbda4fa3bc99d60117c/c/loop-invgen/apache-get-tag.i.p%2Bnlh-reducer.c
  Command-line:
```
scripts/cpa.sh \
  -config config/components/predicateAnalysis-100s-generate-cmc-condition.properties \
  -setprop assumptions.automatonIgnoreAssumptions=true \
  -timelimit 100s
  -spec config/properties/unreach-call.prp \
  ../sv-benchmarks/c/loop-invgen/apache-get-tag.i.p+nlh-reducer.c
```
Produced a condition automaton with invalid match-statements.

Example to reproduce persisting issue on switch-case:
  Task: test/programs/simple/multiple-scopes.c
  Command-line:
```
scripts/cpa.sh \
  -valueAnalysis
  -spec config/properties/unreach-call.prp \
  test/programs/simple/multiple-scopes.c
```
Produces a violation witness with invalid source-code guard.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@33527 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c