Use the full raw signature of a condition for assume edges.
Not just the actual condition of the assume edge.
This reflects the existing behavior for statements,
as well as the JavaDoc of the CFAEdge interface:
Even if there are side effects or post-processing, the **raw** signature
still is the original statement that the CFA edge was created from.

For example, if we have statement 'a++', CPAchecker will split this
into individual statements similiar to 'tmp = a; a; a = tmp + 1'.
Still, all three statements will have raw signature 'a++'.
We also adapt this for assume edges.

This potentially breaks the way assumption automata are built, though:
A condition 'if (p || q)' will be split into two assume
edges [p] and [q] by CPAchecker, and the AssumptionCollectorAlgorithm
will write directions for [p], [not p], [q] and [not q].
When writing these, it uses the raw signature, so we will have edges
in the automaton for '(p || q)' twice, and for '(not (p || q))' twice.
Since the behavior is consistent for input and output, this still
works in theory, but experiments are pending.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@33538 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c