Fix problem with gotos by removing class LoopEntry
We actually do not need to track the exact LoopEntry.
If we only increase the loop counter once we encounter the
node again via which we entered, this can lead to infinite
paths where the counter does not increase. Instead, we now
simply increase the counter whenever we pass one of the
heads of the loop. See issue #708.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@33544 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c