Replace one usage of ProverEnvironmentWithFallback with ProverEnvironment
Inspired by discussion at
https://gitlab.com/sosy-lab/software/cpachecker/-/commit/8d60de88fba8db2916f9649c92ab8e397a824e0e

In this commit, there are mostly changes of parameter types
from ProverEnvironmentWithFallback to a super interface of it
(which do not have any runtime effect)
and the single change of calling solver.newProverEnvironment
instead of new ProverEnvironmentWithFallback.
However, the latter should behave exactly the same as the former
as long as GENERATE_UNSAT_CORE is not requested (which it is not here).

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@33547 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c