ASTConverter: Fix null pointer in method isMainArgumentArray
System.out resulted in a null pointer as System is part of java.lang. As java.lang is automatically imported by Java, it is not in the AST and references to classes of this package cause many problems. This is a bandage solution and might need to be fixed later.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/java-development@33571 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c