NameProvider: Method getObjectScope now returns null if accessing public static field of non declared class
We had a simple test class with public static field

class A {
  public static int i = 1;
};

and a reference A.i. Class a was never instantiated. Method getObjectScope tried to find the declaration of class A and run methods on that declaration. But as A was never declared, we ran into a null pointer.

We now simply return null if field is public static and class was never instantiated. The method returned null in some cases before, so everything should be fine.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/java-development@33572 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c