Remove LDDCPA
This CPA was implemented in 2012 and never changed since then
(just general maintenance commits).
Since 2014 it was marked as unmaintained and produced a warning when
used.
The code needs a binary library libJLDD.so,
which was never added to the repo in a 64-bit version.
Furthermore, in 2015 the 32-bit binary of libJLDD.so was removed,
so since then it was not possible to use the LDDCPA anymore
without manually searching and building the library.

However, there was no link added to the documentation for JLDD
and when doing a web search I could not find any website
or download possibility for this project,
so it seems it is not possible for example to create a new binary of
libJLDD.so.

And there was never a proper license declaration for libJLDD.so
and the (still existing) edu.cmu.sei.rtss.jldd.jar,
so we should remove the latter.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@33663 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c